Extract MediaPlayer from PlaylistToolbar
